refactor: rename app/tenant.py to app/core/authorization.py
Disambiguates from the unrelated app/middleware/tenant.py (a different
concern: tenant *resolution* middleware vs. per-request *authorization*
helpers). Pure rename plus import-path updates across 8 call sites;
no behavior change. First step of scaffolding app/core/ as the
cross-cutting "project layer" for the Django-style per-app restructure.
