+++ app/blueprints/trash.py
@@ -16,7 +16,7 @@ from flask_login import login_required, current_user
 from app.extensions import db
 from app.models import Course, Quiz, Assignment, File, Link, ContentFolder
 from app.helpers import get_jakarta_now
-from app.tenant import get_school_id_or_abort, verify_course_in_school
+from app.core.authorization import get_school_id_or_abort, verify_course_in_school
 
 logger = logging.getLogger(__name__)
